requirements.txt
# Core Framework Dependencies
crewai>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0
neo4j>=5.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Graph Analysis
networkx>=3.0
pyvis>=0.3.2

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Text Processing
unidecode>=1.3.0

# PDF Export
reportlab>=4.0.0

# Note: logging is part of Python's standard library, no need to install

